+++ spec/fixtures/08_assert_nil/in.rb
@@ -1,3 +1,4 @@
-test "there is no kiwi, because they are gross" do
+test "bananas are delicious, kiwis are gross" do
+  assert_not_nil banana
   assert_nil kiwi
 end