+++ spec/fixtures/08_assert_nil/out.rb
@@ -1,1 +1,4 @@
-it "there is no kiwi, because they are gross" { expect(kiwi).to be_nil }
+it "bananas are delicious, kiwis are gross" do
+  expect(banana).to_not be_nil
+  expect(kiwi).to be_nil
+end