+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Unreleased
 ### Added
 - Converts assert_not_nil
 - CLI: creates `target_file` directory if it does not exist
+- Experimental
+  - mocah: with
 
 ### Fixed
 - `__FILE__` keyword in input